backend/requirements.txt
Django
djangorestframework
django-cors-headers
xgboost
scikit-learn
joblib
numpy
pandas
requests
gunicorn
dill